==> 03_Y-Intercept_SSN_Model.Rmd <==
---
title: "ssn_vdeq_tutdeu_comparison"
author: "Michael G. McManus, US EPA/ORD, and Travis Linscome-Hatfield, Neptune and Company"
date: "04/03/2025"
output: html_document
editor_options: 
  chunk_output_type: console
---
## Outline

The purpose of this script is to compare 15 spatial autocovariances for a y-intercept only model. From this comparison, we look at three model comparison values. The model with the smallest Akaike's Information Criterion (AICc), Root Mean Square Prediction Error (RMSPE), and highest squared correlation (cor2) between predicted and true values will subsequently be used in the VDEQ SSN models with covariates in script 05_MLR_SSN_Models. The glances function gives the AICc values, with smaller values indicating better model fit. The loocv (leave-one-out cross validtion) function gives RMSPE and cor2 output, which describe model performance, aka prediction. The table of the 15 model comparisons was saved on 02/10/2025 as a csv file to the outputs folder. The td_exp_eu_exp model was selected.

This approach is saying "space is so important, I am going to model the residual spatial autocorrelation without any covariates". Based on the y-intercept only SSN model, I will compare the forms, Euclidean, Tail Up, and Tail Down, and shapes of the autocovariances. See SSN2 vignette section 4.1.1 Model Summaries about different ways to compare SSN models.

Spatial stream network models typically rely on two families of autocovariance functions: the tail-up and tail-down models. Both autocovariance functions leverage flow-unconnected and flow-connected distances to represent flow-based relationships. Tail-up autocovariance functions capture passive downstream transport (e.g, sediment), whereas tail-down autocovariance functions capture upstream transport (e.g., fish swimming upstream). We can also include the more traditional Euclidean autocovariance functions that are based on straight-line distance, useful for captuing dependence in processes separate from the steam network (e.g., air temperature).
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(SSN2)
library(tidyverse)
library(janitor)

sessionInfo()

# attached base packages:
# [1] stats     graphics  grDevices utils     datasets  methods   base     
# 
# other attached packages:
#  [1] janitor_2.2.0   lubridate_1.9.3 forcats_1.0.0   stringr_1.5.1   dplyr_1.1.4    
#  [6] purrr_1.0.2     readr_2.1.5     tidyr_1.3.1     tibble_3.2.1    ggplot2_3.5.1  
# [11] tidyverse_2.0.0 SSN2_0.2.1  



j_ssn1a <- SSN2::ssn_import("ssn_object/James_071024_pluspreds.ssn", predpts = "sites")
class(j_ssn1a)

names(j_ssn1a)
names(j_ssn1a$obs)

# for consistency I am going to pull the obs from the ssn and apply clean_names so all variable names are in lower case

DFobs <- SSN2::ssn_get_data(j_ssn1a)
DFobs <- clean_names(DFobs)
names(DFobs)

# this shows that DFobs is both an sf and data.frame
class(DFobs)

# note ssn_put_data requires sf object and SSN2 object
# this is putting cleaned names of DFobs back into SSN2 object
j_ssn1a <-  SSN2::ssn_put_data(DFobs,j_ssn1a)

```

# 1.0 One Way:  SSN Y-Intercept
All these models have the same fixed effects structure so REML estimation, the default, can be used. See the last paragraph in SSN2 vignette section 4.1 Linear SSN Models for why restricted maximum likelihood (REML) is the default and when maximum likelihood (ML) should be used. For the one-way models, the notation assigned the models below gives the two-letter spatial autocovariance forms first as tu, td, or eu for tail up, tail down, or Euclidean, respectively, followed by an underscore, and then three letters describing the shape, exp, lin, or gau for exponential, linear, or Gaussian, respectively.

The <https://pet221.github.io/SSNbler/articles/introduction.html#calculating-additive-function-values-afvs> gives an explanation of the additive function values, which are specified with the additive argument for the ssn_lm function. Basically, the additive function value is what allows spatial weighting to be assigned at confluence where, for example, a large mainstem converges with a smaller tributary. In which case, the large mainstem would be given more weighting than the small tributary.


```{r oneway_ssn_yintercept}
tu_exp <- ssn_lm(
  formula = vscivcpmi ~ 1,
  ssn.object = j_ssn1a,
  tailup_type = "exponential",
  taildown_type = "none",
  euclid_type = "gaussian",
  additive = "afv_area"
)
summary(tu_exp)
varcomp(tu_exp)

# plot(tu_exp, which = c(1:6))

tu_exp_res <- as.data.frame(tu_exp$residuals)

ggplot(tu_exp_res, aes(x=standardized)) + geom_histogram()
ggplot(tu_exp_res, aes(x=standardized)) + geom_boxplot()
ggplot(tu_exp_res, aes(x=standardized)) + geom_density()
# maximum stream and Euclidean distances
tu_exp$euclid_max # this matches my Euclid max
tu_exp$tail_max # this is much larger than my stream max




tu_lin <- ssn_lm(
  formula = vscivcpmi ~ 1,
  ssn.object = j_ssn1a,
  tailup_type = "linear",
  taildown_type = "none",
  euclid_type = "none",
  additive = "afv_area"
)

td_exp <- ssn_lm(
  formula = vscivcpmi ~ 1,
  ssn.object = j_ssn1a,
  tailup_type = "none",
  taildown_type = "exponential",
  euclid_type = "none",
  additive = "afv_area"
)

td_lin <- ssn_lm(
  formula = vscivcpmi ~ 1,
  ssn.object = j_ssn1a,
  tailup_type = "none",
  taildown_type = "linear",
  euclid_type = "none",
  additive = "afv_area"
)

eu_exp <- ssn_lm(
  formula = vscivcpmi ~ 1,
  ssn.object = j_ssn1a,
  tailup_type = "none",
  taildown_type = "none",
  euclid_type = "exponential",
  additive = "afv_area"
)

eu_gau <- ssn_lm(
  formula = vscivcpmi ~ 1,
  ssn.object = j_ssn1a,
  tailup_type = "none",
  taildown_type = "none",
  euclid_type = "gaussian",
  additive = "afv_area"
)

models_yintercept <- glances(tu_exp, tu_lin, td_exp, td_lin, eu_exp, eu_gau)
loocv(tu_exp)
loocv(tu_lin)
loocv(td_exp)
loocv(td_lin)
loocv(eu_exp)
loocv(eu_gau)
```

# 2.0 Two Way:  SSN Y-Intercept
```{r twoway_ssn_yintercept}
tu_exp_td_exp <- ssn_lm(
  formula = vscivcpmi ~ 1,
  ssn.object = j_ssn1a,
  tailup_type = "exponential",
  taildown_type = "exponential",
  euclid_type = "gaussian",
  additive = "afv_area"
)
summary(tu_exp_td_exp)
varcomp(tu_exp_td_exp)
tu_exp_td_exp_res <- as.data.frame(tu_exp_td_exp$residuals)

ggplot(tu_exp_td_exp_res, aes(x=standardized)) + geom_histogram()
ggplot(tu_exp_td_exp_res, aes(x=standardized)) + geom_boxplot()
ggplot(tu_exp_td_exp_res, aes(x=standardized)) + geom_density()
# maximum stream and Euclidean distances
tu_exp_td_exp$euclid_max # this matches my Euclid max
tu_exp_td_exp$tail_max # this is much larger than my stream max

tu_exp_td_lin <- ssn_lm(
  formula = vscivcpmi ~ 1,
  ssn.object = j_ssn1a,
  tailup_type = "exponential",
  taildown_type = "linear",
  euclid_type = "none",
  additive = "afv_area"
)

tu_lin_td_exp <- ssn_lm(
  formula = vscivcpmi ~ 1,
  ssn.object = j_ssn1a,
  tailup_type = "linear",
  taildown_type = "exponential",
  euclid_type = "none",
  additive = "afv_area"
)

tu_lin_td_lin <- ssn_lm(
  formula = vscivcpmi ~ 1,
  ssn.object = j_ssn1a,
  tailup_type = "linear",
  taildown_type = "linear",
  euclid_type = "none",
  additive = "afv_area"
)

td_exp_eu_exp <- ssn_lm(
  formula = vscivcpmi ~ 1,
  ssn.object = j_ssn1a,
  tailup_type = "none",
  taildown_type = "exponential",
  euclid_type = "exponential",
  additive = "afv_area"
)

summary(td_exp_eu_exp)
varcomp(td_exp_eu_exp)

td_exp_eu_gau <- ssn_lm(
  formula = vscivcpmi ~ 1,
  ssn.object = j_ssn1a,
  tailup_type = "none",
  taildown_type = "exponential",
  euclid_type = "gaussian",
  additive = "afv_area"
)


td_lin_eu_exp <- ssn_lm(
  formula = vscivcpmi ~ 1,
  ssn.object = j_ssn1a,
  tailup_type = "none",
  taildown_type = "linear",
  euclid_type = "exponential",
  additive = "afv_area"
)

td_lin_eu_gau <- ssn_lm(
  formula = vscivcpmi ~ 1,
  ssn.object = j_ssn1a,
  tailup_type = "none",
  taildown_type = "linear",
  euclid_type = "gaussian",
  additive = "afv_area"
)

models_yintercept <- glances(tu_exp, tu_lin, td_exp, td_lin, eu_exp, eu_gau, tu_exp_td_exp, tu_exp_td_lin, tu_lin_td_exp, tu_lin_td_lin, td_exp_eu_exp, td_exp_eu_gau, td_lin_eu_exp, td_lin_eu_gau)

loocv(tu_exp_td_exp)
loocv(tu_exp_td_lin)
loocv(tu_lin_td_exp)
loocv(tu_lin_td_lin)
loocv(td_exp_eu_exp)
loocv(td_exp_eu_gau)
loocv(td_lin_eu_exp)
loocv(td_lin_eu_gau)

```

# 3.0 Threeway:  SSN Y-Intercept
```{r threeway_ssn_yintercept}
tu_td_eu_exp <- ssn_lm(
  formula = vscivcpmi ~ 1,
  ssn.object = j_ssn1a,
  tailup_type = "exponential",
  taildown_type = "exponential",
  euclid_type = "exponential",
  additive = "afv_area"
)
summary(tu_td_eu_exp)
varcomp(tu_td_eu_exp)
tu_td_eu_exp_res <- as.data.frame(tu_exp_td_exp$residuals)

ggplot(tu_td_eu_exp_res, aes(x=standardized)) + geom_histogram()
ggplot(tu_td_eu_exp_res, aes(x=standardized)) + geom_boxplot()
ggplot(tu_td_eu_exp_res, aes(x=standardized)) + geom_density()
# maximum stream and Euclidean distances
tu_td_eu_exp$euclid_max # this matches my Euclid max
tu_td_eu_exp$tail_max 

models_yintercept <- glances(tu_exp, tu_lin, td_exp, td_lin, eu_exp, eu_gau, tu_exp_td_exp, tu_exp_td_lin, tu_lin_td_exp, tu_lin_td_lin, td_exp_eu_exp, td_exp_eu_gau, td_lin_eu_exp, td_lin_eu_gau, tu_td_eu_exp)

View(models_yintercept)

loocv(tu_td_eu_exp)

write.csv(models_yintercept, file="outputs/models_yintercept.csv",sep=",",row.names=F)
```
